Centralize every monkeypatch in torch_mojo_backend/monkeypatching.py
Move the seven PyTorch patches out of mojo_device/register.py (accelerator
synchronize/stream API, torch.Stream/torch.Event dispatch, the foreach
allowlists, HANDLED_TYPES / FakeTensorMode / FunctionalTensor tracing
gates, dynamo's VariableBuilder table, FakeTensor.__new__) and the one
internal patch from apple_optimizations.py into a single module. Each
patch is one function whose docstring says what upstream lacks, so it can
be turned into a PyTorch PR and deleted here; register_mojo_devices calls
apply_torch_monkeypatches once, right after the PrivateUse1 backend is set
up. Behaviour is unchanged.

A CPU-only test scans the package with ast and fails on any assignment to
a torch-rooted attribute outside that file, and AGENTS.md states the rule.
