+++ torch_mojo_backend/mojo_device/apple_optimizations.py
@@ -1,17 +1,7 @@
+from torch_mojo_backend.monkeypatching import (
+    use_apple_fast_add as _enable_apple_fast_add,
+)
 from torch_mojo_backend.torch_compile_backend.utils import get_accelerators
-
-
-def _enable_apple_fast_add():
-    """Select the isolated Metal contiguous-add implementation.
-
-    Patching once during device registration keeps CUDA and ROCm on the
-    original Python and Mojo paths with no per-call target check.
-    """
-    from torch_mojo_backend.eager_kernels import aten_fast
-
-    # Deliberate module-attribute monkeypatch; ty compares the two
-    # same-signature `def`s nominally, so this can never structurally match.
-    aten_fast.fast_aten_add = aten_fast.fast_aten_add_apple  # ty: ignore[invalid-assignment]
 
 
 def register_apple_optimizations():
